Extract HTML and text manipulation methods into HtmlTextUtils module
