+++ app/services/normalizer/base.rb
@@ -7,6 +7,7 @@
 #
 module Normalizer
   class Base
+    include HtmlTextUtils
     # @param feed_entry [FeedEntry] the feed entry to normalize
     def initialize(feed_entry)
       @feed_entry = feed_entry
@@ -88,26 +89,6 @@ module Normalizer
 
     def raw_data
       feed_entry.raw_data
-    end
-
-    def strip_html(text)
-      return "" if text.blank?
-
-      doc = Nokogiri::HTML::DocumentFragment.parse(text)
-      doc.text.strip.gsub(/\s+/, " ")
-    end
-
-    def extract_images_from_content(content)
-      return [] if content.blank?
-
-      doc = Nokogiri::HTML::DocumentFragment.parse(content)
-      doc.css("img").map { |img| img["src"] }.compact
-    end
-
-    def truncate_content(content)
-      return content if content.length <= Post::MAX_CONTENT_LENGTH
-
-      content.truncate(Post::MAX_CONTENT_LENGTH, separator: " ")
     end
 
     def normalize_source_url(url)
